# api/v2/openapi.yaml
openapi: 3.1.0
servers:
  - url: /api/v2
info:
  title: Projects Service
  version: 0.0.1
paths:
...
  "/get-project":
    post:
      summary: Get project
      description: Get project
      operationId: getProject
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GetProjectRequest"
        required: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetProject"
        '404':
          description: Not found
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
...
  "/add-project":
    post:
      summary: Upsert a project
      description: Upsert a project
      operationId: addProject
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddProjectRequest'
        required: true
      responses:
        '200':
          description: Success
        default:
          $ref: "#/components/responses/Error"
...
security:
  - apiKey: []
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-Api-Key
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        code:
          type: integer
          format: int64
        message:
          type: string
      required:
        - code
        - message
...
    GetProjectRequest:
      type: object
      properties:
        identifier:
          $ref: "#/components/schemas/Identifier"
      required: [identifier]
    GetProject:
      type: object
      properties:
        project:
          $ref: "#/components/schemas/Project"
      required: [project]
...
    AddProjectRequest:
      type: object
      properties:
        project:
          type: object
          properties:
            identifiers:
              type: array
              items:
                $ref: "#/components/schemas/Identifier"
            names:
              type: array
              items:
                $ref: "#/components/schemas/Text"
            descriptions:
              type: array
              items:
                $ref: "#/components/schemas/Text"
            startDate:
              type: string
            endDate:
              type: string
            deleted:
              type: boolean
            attributes:
              type: array
              items:
                $ref: "#/components/schemas/Attribute"
          required: [identifiers]
      required: [project]
    Project:
      type: object
      properties:
        identifiers:
          type: array
          items:
            $ref: "#/components/schemas/Identifier"
        names:
          type: array
          items:
            $ref: "#/components/schemas/Text"
        descriptions:
          type: array
          items:
            $ref: "#/components/schemas/Text"
        startDate:
          type: string
        endDate:
          type: string
        deleted:
          type: boolean
        attributes:
          type: array
          items:
            $ref: "#/components/schemas/Attribute"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [identifiers, deleted, createdAt, updatedAt]
...
    Attribute:
      type: object
      properties:
        scope:
          type: string
          minLength: 1
        key:
          type: string
          minLength: 1
        value:
          type: string
          minLength: 1
      required: [scope, key, value]
    Identifier:
      type: object
      properties:
        kind:
          type: string
          minLength: 1
        value:
          type: string
          minLength: 1
      required: [kind, value]
...
    Text:
      type: object
      properties:
        lang:
          type: string
          minLength: 1
        value:
          type: string
          minLength: 1
      required: [lang, value]